Segmentation_Satellit_Image/requirements.txt
numpy==1.22.0
patchify==0.2.3
Pillow==10.2.0
plotly==5.9.0
scikit_learn==0.24.0
segmentation_models==1.0.1
streamlit==1.28.1
tensorflow==2.13.0rc0
protobuf==3.20.*
opencv-python==4.5.3.56
matplotlib==3.7.1
tqdm==4.62.3
keras==2.13.1






